printSelectMenu 수정 도전 과제 1레벨에 요구하는 기능 추가 및 level4에서 복사해와서 import level4  되어있던거 삭제

+++ src/main/java/level6/MenuItem.java
@@ -25,7 +25,7 @@ public class MenuItem {
         return price;
     }
 
-    public static void totalPrice(int price){   //선택한 메뉴 가격 더하기
+    public static void totalPrice(long price){   //선택한 메뉴 가격 더하기
         totalPrice += price;
     }
